Replace jackson-databind with jodd-json
Replaces jackson-databind with jodd-json[0].

The motivation behind this is to reduce the likeliness of
a version conflict in existing applications. Jackson is
one of the most common JSON libraries out there, but it has
some severe issues when multiple versions are available on the
classpath.

jodd-json is a small JSON parser/serializer library with a minimal
footprint. The resulting shaded jar of java-jwt is about half the size
compared to when its bundled with jackson-databind.

The patch has a bad side effect in that it created different tokens
than the jackson-bind implementation, since it escapes all the
characters defined in RFC4627, while jackson-databind only escapes
the MUST characters. This is visible in the test cases that include
"/" in the claim values.

Another "bad" effect is the use of Map<String, Object>, and Object
references. I've tried to add the needed type checks, but there
might dragons left to slay.

This commit should fix issue #18 and similar multiple jackson versions
issues that might be out there is the wild.

[0]: http://jodd.org/doc/json/