Support custom validation
Adds the ability for users to add one or more custom validations to the
JWT verification process, by using `withCustomValidation` on the
validation builder.

This is useful to workaround the inability to perform particular
validations directly with the library, such as issue #246, or to do other
validations that are not possible with the built-in validations.